issue #47 feature : 채팅 메시지 레이아웃 내 메시지와 상대 메시지 분리
