+++ app/src/main/res/layout/chat_msg_item_received.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/parent"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    >
+    android:layout_height="wrap_content">
 
     <Space
         android:id="@+id/space_header"
@@ -25,7 +23,7 @@
         android:layout_marginRight="5dp"
         android:text="username"
         app:layout_constraintTop_toBottomOf="@id/space_header"
-        tools:ignore="MissingConstraints"
+        app:layout_constraintStart_toStartOf="parent"
         />
 
 
@@ -35,7 +33,7 @@
         android:layout_height="wrap_content"
         android:layout_marginLeft="5dp"
         android:layout_marginRight="5dp"
-        android:background="@drawable/chat_msg_bubble_right"
+        android:background="@drawable/chat_msg_bubble_left"
         android:paddingLeft="10dp"
         android:paddingTop="8dp"
         android:paddingRight="10dp"
@@ -47,7 +45,7 @@
         app:layout_constraintTop_toBottomOf="@id/user_name"
         app:layout_constraintWidth_max="wrap"
         app:layout_constraintWidth_percent="0.8"
-        tools:ignore="MissingConstraints"
+        app:layout_constraintStart_toStartOf="parent"
         />
 
     <TextView
@@ -59,7 +57,7 @@
         android:text="time"
         app:layout_constrainedWidth="true"
         app:layout_constraintBottom_toBottomOf="@+id/msg"
-        tools:ignore="MissingConstraints"
+        app:layout_constraintStart_toEndOf="@id/msg"
         />
